Rename XxxStreamProperties files to XxxStreamProperty
* New files are:
     * VideoStreamProperty
     * AudioStreamProperty
     * DataStreamProperty
     * SubtitleStreamProperty
     * AttachementStreamProperty
     * UnknownStreamProperty
